+++ src/misc/sound_media.h
@@ -9,6 +9,7 @@
 //extern Sound hit_sound;
 extern Sound general_sounds[5];
 
+
 //function to load general audio
 bool LoadGeneralAudio();
 
@@ -16,7 +17,7 @@ bool LoadGeneralAudio();
 void PlayGeneralSound(GeneralSoundID& sound_id);
 
 //function to play character sound based on character sound id and character
-void PlayCharacterSound(std::uint8_t& char_index, CharSoundID& sound_id);
+void PlayCharacterSound(std::int16_t& char_index, CharSoundID& sound_id);
 
 //function to unload general audio
 void UnloadGeneralAudio();
